+++ pdftotree/TreeExtract.py
@@ -1,12 +1,17 @@
 import html
 import logging
+import os
 from functools import cmp_to_key
 from typing import Any, Dict, List, Tuple
 from xml.dom.minidom import Document, Element
 
 import numpy as np
 import tabula
-from pdfminer.layout import LTChar, LTTextLine
+from pdfminer.layout import LAParams, LTChar, LTPage, LTTextLine
+from pdfminer.pdfdocument import PDFDocument
+from pdfminer.pdfinterp import PDFPageInterpreter, PDFResourceManager
+from pdfminer.pdfpage import PDFPage
+from pdfminer.pdfparser import PDFParser
 from pdfminer.utils import Plane
 
 from pdftotree._version import __version__
@@ -21,7 +26,7 @@ from pdftotree.utils.lines_utils import (
     reorder_lines,
 )
 from pdftotree.utils.pdf.pdf_parsers import parse_layout, parse_tree_structure
-from pdftotree.utils.pdf.pdf_utils import PDFElems, analyze_pages, normalize_pdf
+from pdftotree.utils.pdf.pdf_utils import CustomPDFPageAggregator, PDFElems
 from pdftotree.utils.pdf.vector_utils import column_order, reading_order
 
 
@@ -35,12 +40,6 @@ class TreeExtractor(object):
         self.pdf_file = pdf_file
         self.elems: Dict[int, PDFElems] = {}  # key represents page_num
         self.font_stats: Dict[int, Any] = {}  # key represents page_num
-        self.lines_bboxes = []
-        self.alignments_bboxes = []
-        self.intersection_bboxes = []
-        self.bboxes = []
-        self.candidates = []
-        self.features = []
         self.iou_thresh = 0.8
         self.scanned = False
         self.tree: Dict[int, Any] = {}  # key represents page_num
@@ -88,9 +87,39 @@ class TreeExtractor(object):
     def parse(self):
         is_scanned = False
         lin_seg_present = False
-        for page_num, layout in enumerate(analyze_pages(self.pdf_file)):
+        layouts: List[LTPage] = []
+
+        log = logging.getLogger(__name__)
+        # Open a PDF file.
+        with open(os.path.realpath(self.pdf_file), "rb") as fp:
+            # Create a PDF parser object associated with the file object.
+            parser = PDFParser(fp)
+            # Create a PDF document object that stores the document structure.
+            # Supply the password for initialization.
+            document = PDFDocument(parser, password="")
+            # Create a PDF resource manager object that stores shared resources.
+            rsrcmgr = PDFResourceManager()
+            # Set parameters for analysis.
+            laparams = LAParams(char_margin=1.0, word_margin=0.1, detect_vertical=True)
+            # Create a PDF page aggregator object.
+            device = CustomPDFPageAggregator(rsrcmgr, laparams=laparams)
+            # Create a PDF interpreter object.
+            interpreter = PDFPageInterpreter(rsrcmgr, device)
+            # Process each page contained in the document.
+            for page_num, page in enumerate(PDFPage.create_pages(document)):
+                try:
+                    interpreter.process_page(page)
+                except OverflowError as oe:
+                    log.exception(
+                        "{}, skipping page {} of {}".format(oe, page_num, self.pdf_file)
+                    )
+                    continue
+                layout = device.get_result()
+                layouts.append(layout)
+
+        for page_num, layout in enumerate(layouts):
             page_num += 1  # indexes start at 1
-            elems, font_stat = normalize_pdf(layout, scaler=1)
+            elems, font_stat = device.normalize_pdf(layout, scaler=1)
             self.elems[page_num] = elems
             self.font_stats[page_num] = font_stat
             # code to detect if the page is scanned
@@ -272,8 +301,11 @@ class TreeExtractor(object):
             page = doc.createElement("div")
             page.setAttribute("class", "ocr_page")
             page.setAttribute("id", f"page_{page_num}")
+            width = int(self.elems[page_num].layout.width)
+            height = int(self.elems[page_num].layout.height)
             page.setAttribute(
-                "title", f"bbox 0 0 {int(pwidth)} {int(pheight)}; ppageno {page_num-1}"
+                "title",
+                f"bbox 0 0 {width} {height}; ppageno {page_num-1}",
             )
             body.appendChild(page)
             # TODO: We need to detect columns and sort acccordingly.